remove AuthToken, handle in controller
